Move ScrollLinker from util file to utils/ScrollLinker
Caleydo/cohort#693
